Added the unsubscribe email logic (need to create actual send email logic)

+++ src/utils/blitz-utils.ts
@@ -60,12 +60,15 @@ export const formatZodError = (err: any) => {
   });
 };
 
-const EMAIL_VERIFY_LINK_IN_HOURS = 4;
-
-const createToken = async ({ userId, userEmail, tokenType }) => {
+const createToken = async ({
+  userId,
+  userEmail,
+  tokenType,
+  expiryHours = 4,
+}) => {
   const token = generateToken();
   const hashedToken = hash256(token);
-  const expiresAt = addHours(new Date(), EMAIL_VERIFY_LINK_IN_HOURS);
+  const expiresAt = addHours(new Date(), expiryHours);
 
   await db.token.create({
     data: {
@@ -83,19 +86,26 @@ export const regenerateToken = async ({
   userId,
   userEmail,
   tokenType,
+  expiryHours = 4,
+  deleteExisting = true,
 }: {
   userId: string;
   userEmail: string;
   tokenType: TokenType;
+  expiryHours?: number;
+  deleteExisting?: boolean;
 }): Promise<string> => {
-  await db.token.deleteMany({
-    where: { type: tokenType, userId },
-  });
+  if (deleteExisting) {
+    await db.token.deleteMany({
+      where: { type: tokenType, userId },
+    });
+  }
 
   const token = await createToken({
     userId,
     userEmail,
     tokenType,
+    expiryHours,
   });
 
   return token;
